requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
sqlalchemy>=1.4.0
httpx>=0.23.0
requests>=2.31.0
python-dotenv>=0.19.0
pydantic>=1.8.0
python-multipart>=0.0.5
tabulate
colorama
gradio